requirements.txt
flask
flask_sqlalchemy
werkzeug
pypdf
python-dotenv
youtube-transcript-api
yt-dlp
groq
feedparser
gunicorn